pyproject.toml
[project]
name = "airflow-providers-navitia"
version = "0.1.0"
description = "Navitia hook and operator for Airflow"
authors = [
    {name = "Jonathan Perron",email = "jonathan@perron.bzh"}
]
license = {text = "MIT license"}
requires-python = ">=3.9"
dependencies = [
    "apache-airflow>=2.10",
    "apache-airflow-providers-postgres>=6.0.0, <7",
    "python-navitia-client>1.1.5"
]
classifiers = [
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.urls]
Source = "https://github.com/jonperron/airflow-provider-navitia"

[project.readme]
file = "README.md"
content-type = "text/markdown"
